Add the Bevy game engine (#3)
Since we are building a video game, we want to use a game engine for
the project. Game engines implement common tasks and functionality, for
example rendering pipelines and audio support. Working with a game
engine means that we won't have to implement these features ourselves,
saving us time and effort.

Because the video game is targeted at programmers, it must feature an
API that enables players to interact with the simulation. This requires
a tight integration between the engine and the API. This is certainly
easiest if the engine and API are implemented using the same
programming language so that resources can be shared between the two.
Previous experiments have shown that a gRPC API can be integrated with
the Bevy game engine, which is why we have chosen it for this project.

Bevy has been added as a dependency, and an empty app is launched when
the game is run.