Prototype systems that render the routing grid (#4)
The first systems inside Bevy have been implemented. The systems set up
a 2D camera, and render the routing grid as well as an airport.

Abstractions around positions on the game's 2D plane have been created
to make it easier to place entities on the map. Coordinates are simply
a point on the plane, while tiles map to a uniform grid that is used
for routing as well as rendering.

The implementation relies heavily on hard-coded parameters to make
prototyping easier. The window size is fixed, and the size of the map
is derived from it. This eliminates the need to have a movable camera.
The airport is not yet a resource in the simulation, but is instead
hard-coded to be at the center of the map (x: 0, y: 0). Working with
these assumption makes it possible to introduce new systems and
complexities gradually, without the need to manage everything inside
the simulation yet.