Prototype system that follows flight plans (#5)
A new system has been prototyped that moves a plane along its flight
plan. The plane passed each routing node in the flight plan until it
reaches the destination.

Flight plans are stored in reverse order so that nodes can be easily
removed from the end of the vector when the plane passes them. This
convention needs to be checked and enforced in a future commit.