Create system that despawns airplanes
A new system has been added that checks if an airplane has reached its
destination, and despawns it if it did. The system deliberately does not
check if the airplane has reached the airport, but instead checks if the
flight plan has any nodes left. This creates flexibility for future game
mechanics, for example random planes crossing the map.