Prototype path finding for airplanes (#7)
The simulation of airplanes has been extended with a path finding
algorithm to create the flight plans. This makes it possible to spawn
airplanes at random locations around the map, and have them fly to the
airport where they get despawned.

The implementation of the path finding algorithm is very naive, and
simply finds a suitable path to the airport. Since it will ultimately
be the task of the user to provide the flight plans, we might not need
a path finding algorithm at all in the future.