Create workspace with existing crate as member
The project structure has been reorganized in preparation for the API. A
Rust workspace has been created for the project, and the existing code
has been moved into the `atc-game` crate. This makes room to introduce
an `atc` crate with the API layer.

Keeping the game and its API in the same repository increases the
iteration speed during prototyping. But in the long-term, we might want
to revert this decision and move the API into its own repository for the
following reasons:

  1. Splitting the code into separate repositories allows us to open
     source the API layer while keeping the game itself closed source.
     This might be relevant if we ever want to sell the game.
  2. The API will have client libraries for many different languages,
     which need to be built whenever the Protocol Buffers change.
     Keeping this code in a separate repository makes it easier to set
     up the different build systems.

Depending on the future development of the project, neither might be a
good enough argument to split the repository again.