Create event service (#10)
The event service in the API specification has been implemented. The
service listens to an event bus, and streams all events to connected
clients.

The event bus is implemented using a broadcast channel from tokio. This
type of channel has been chosen, since the event bus has many producers
(different systems in the game) and many consumers (event service and a
data store).