Crate airplane service (#11)
A new service has been added to the API that enables users to query the
airplanes. When a request is received, the service uses the provided id
to look for the airplane in a store. If an airplane is found, it is
returned. Otherwise, a gRPC status of NOT_FOUND is returned.

The store is implemented as a concurrent hash map, which is provided by
the dashmap crate. A task is spawned on the async runtime to watch the
event bus, and to update the airplanes in the store based on the events
on the bus. The DashMap handles concurrency internally, making it easy
to integrate the data structure in the game.