Validate flight plans (#14)
The flight plan component has been extended with validation logic. When
a plane is spawned, a default flight plan will be created for it. When
the user then updates the flight plan, certain rules must be fulfilled.
Most importantly, the flight plan must start with the node that the
plane is currently heading towards.