Create API to update flight plans (#12)
A new API endpoint has been created that enables users to update the
flight plan of an airplane. The new flight plan gets validated, and
any errors are reported back to the user. If the plan is valid, a
command is queued to update it and a new system picks up the plan and
applies it.