Create app states and an API to start a game (#16)
The game can now be in different states, which unblocks further
development of the simulation. The default state is the `ready` state,
which shows a menu screen to the user. The screen has only been mocked
yet, and does not have its final functionality. The player can send a
request to start a game, which transitions from the `ready` state to
the `running` state. This state had previously been implemented, and
runs the simulation.

Future work can now implement failure conditions, which transition the
game from the `running` state back to the `ready` state.