Reduce screen size
The screen size has been reduced to be compatible with our streaming
setup. We are currently streaming parts of the development at 720p, and
are experimenting with different approaches to downscale the display
resolution to the stream. The original size of the game was higher than
the stream, which caused issues with the latest attempt to produce a
sharp stream.

This is only a temporary change during development, and will have to be
reconsidered once we start iterating on the final map size.