Measure test coverage (#24)
The test coverage is now collected during CI runs using the tool
tarpaulin. Since tarpaulin only runs on Linux, it has not been
integrated in the local development environment.

Measuring test coverage allows us to identify parts of the code base
that lack tests, and thus might be a source of bugs. Over time and with
regular refactoring sessions, we aim to eventually cover most of the
code base with automated tests.