Group project modules by their function (#27)
The layout of the project has been changed to group the different
modules by their function. This promotes certain modules to the root of
the repository, and makes it easier for users to find what they are
looking for.

The Protocol Buffers have been moved into the `api` directory.
Previously, users were required to have knowledge of the Protocol Buffer
standard to understand what the `protobufs` directory was all about.
Using the generic `api` term now makes it easier to reason about the
contents of the directory.

The prototype Rust client has been placed into the `examples` directory.
This is considered a short-term solution, since the ultimate goal is to
extract the client into its own repository. We can then mark the
repository as a template repository and make it easy for players to use
the starter template for their own bot.

The game is now in the root `game` directory. This hopefully makes it a
bit easier to find the assets for the game, which before were deeply
nested inside the `src` folder.

Client libraries will be placed into the `sdk` directory. These will get
their own build and publish workflows in CI.