Create API to inspect ATC itself
A new API service and endpoint have been created to inspect ATC itself.
The only endpoint of the service returns the version of the game, which
can be used by clients to determine if they are running on the same
version as the server.

The service might be extended in the future to include more information
about the game itself as well as its operation.