Configure automated release notes (#29)
GitHub has a built-in feature to generate release notes from
previous commits. We are evaluating this feature to reduce the
complexity of drafting and publishing a new release.

The feature uses issue labels to group entries in the changelog. A new
configuration file and workflow have been added to manage the labels in
this repository.