Create script that prepares a new release (#31)
A new script that prepares new releases has been added to the
repository. The script updates the manifests of all packages to the new
version, and puts links into the auto-generated changelog from GitHub.
The changes are committed to a release branch, which can be reviewed and
then pushed to GitHub.