Create and add routing grid to the map (#37)
The map has been refactored to be a resource in the game. The map has a
reference to the location of the airport, and a list of nodes that form
the routing grid.

Some parts of the code have been refactored to get the location of the
airport from the map resource. This avoids hard-coded values across the
code base.

The routing grid is used to validate flight plans, and to generate
random routes for new airplanes.

Refactoring the map and introducing the routing grid enabled creating a
"runway" for the airport. The airport is now in a dead-end location and
only reachable through one node. This introduces a certain amount of
realism, but more importantly will allow us to refactor various systems
to be more maintainable.