Refactor store to manage all shared state
The store has been refactored to not only manage airplanes but also the
game state and the map. This makes it the single source of truth for all
shared data. Another benefit is that there is only a single place that
needs to be reset between games now.

Airplanes are still stored in a DashMap, while the game state and the
map are placed inside a shared mutex. Both are mostly read, and only
barely written, so there shouldn't be any performance issues with this
choice.

The store watcher has been refactored as well to take over the
responsibilities of the game state watcher, which in turn has been
removed. This again centralizes all state management in a single
location.