Refactor airplane lifecycle (#40)
The systems that handle the lifecycle of an airplane have been
refactored to increase the maintainability of the code. Especially the
logic around landing and despawning an airplane had crept into many
other systems as well.

A new component has been added that indicates if an airplane is landing.
Airplanes tagged with this component are watched by the restored
despawning system, and removed once they reach the airport. The system
that generates new flight plans ignores any airplane that is landing,
since those airplanes do not need a new flight plan. Since there is no
valid path away from the airport again, the system would also fail to
generate such a plan and crash the game.

Since airplanes move so slow that they cannot cross a tile in a single
frame, the execution order of the systems does not matter anymore. With
the new component, there is no interference between systems and it is
safe to execute them in parallel.