Create type for airports and add it to the map (#43)
The representation of airports in the API and in the game itself has
been changed. Airports have their own struct (internally) and message
type (API) now, which contains a node and a new feature called a tag.
The node represents the location of the airport, and tags will be used
in the future to split airplanes across many airports.

The internal map representation has been updated to use the new airport
struct, but no behavior has been changed yet to make use of the tag.