Add tags to airplanes and airports (#44)
Tags are a new feature in the game, which unlocks the creation of
different airports. Airports and airplanes are spawned with a tag that
has the name of a color, and airplanes can only land at airports with
the same tag.

If an airplane tries to land at the wrong airport, it does a go-around
maneuver. The airplane will fly past the airport and join the routing
grid on the opposite site of the runway. A new event has been added to
inform the player of this.