Add second airport to the map (#46)
The map and the systems working with it have been refactored to support
multiple airports in the game. The map now has a list of airports, both
in its internal representation and in the API message type. The map has
two airports by default, one in blue and one in red. Airplanes are
spawned with either color with a distribution of 50/50.