Relicense project under MIT or Apache-2.0 (#59)
The project is relicensed under the dual license that is common practice
in the Rust ecosystem. This makes it easier for people to build apps
that consume the client libraries of the project, and to contribute to
the public artifacts of the game such as its documentation.

The core of the game is open sourced as well, although we have not yet
decided on a contribution policy. We might only make the game source
available, and not accept any contributions. Please check the main
branch for the most current contribution guide.