Automate the publishing of releases to itch.io (#64)
Initially, the game will be distributed through itch.io [^1], a
platform for indie games. itch.io provides tools that make it easy to
publish and distribute a game to players, which massively reduces the
burden on us to produce builds and installers for the different
platforms.

A new GitHub Action has been created that runs when a new release is
created. Pre-releases are pushed to a beta channel on itch.io, while
regular releases are pushed to the main channel.

Releasing the game for macOS requires us to create an app bundle [^2],
which is done with `cargo-bundle`. `cargo-bundle` can automatically copy
resources into the bundle, but it is missing an API to access them
dynamically from the program [^3]. For this reason, the GitHub Action
moves the resources back so that they are next to the binary.

[^1]: https://jdno.itch.io/auto-traffic-control
[^2]: https://itch.io/docs/itch/integrating/platforms/macos.html
[^3]: https://github.com/burtonageo/cargo-bundle/issues/12