Reorganize GitHub Actions (#65)
The number of GitHub Actions has grown quite a lot since we set up the
project, as has their complexity. Actions are no longer only run when
code is pushed, but also when releases are created. To make it easier to
manage the workflows with their different triggers and conditions, the
files have been reorganized and renamed.

The release workflows also use a different cache compared to normal
builds to avoid mixing debug and release builds.