Postpone NPM release (#76)
The original plan was to release the game with an official client
library in Rust and NPM. Due to unforeseen complications with the NPM
build, we decided to postpone the release of that library. References to
a Node client have been removed from the documentation until the release
is ready.