Release 0.2.0 (#85)
The second release extends the original prototype with a new game
mechanic: Airplanes and airports now have a color that must match to
land the airplane. The APIs of the game have been extended to provide
more information about the map and the different colors. And a public
website with preliminary documentation has been created.