Switch to new variant to access world resources (#96)
The Bevy team noticed that most invocations of `get_resource` were
followed by an immediate `unwrap`. A new API [^1] was introduced that
panics internally, thus avoiding the need for users to unwrap
themselves. The old API still remains for users who want to handle
the `Option` manually.

In our case, we want to panic if the message buses cannot be accessed.
The game cannot work properly without them, and the caller has no way of
fixing the problem.

[^1]: https://bevyengine.org/news/bevy-0-7/#world-resource