Write tutorial for a TypeScript bot
The documentation has been expanded with an in-depth tutorial on how to
play the game with TypeScript. The tutorial covers the installation of
the game through the itch app, setting up a new project for the bot, and
interacting with the API to start a game and subscribe to events. Topics
not covered are path finding and collision detection, which are left as
an exercise for the reader.

The code in the tutorial is based on jdno/node-traffic-controller [^1],
which can be used by players as a template to bootstrap their bot.

[^1]: https://github.com/jdno/node-traffic-controller