Fix the rotation of airplanes (#129)
When airplanes reached the end of their flight plan, some random nodes
were added to the plan to keep the airplane moving. When that happened,
however, the rotation of the airplane's sprite was not updated. This
resulted in airplanes moving sideways for a step in their flight plan,
before being rotated correctly during the next step.

To fix this, a new system has been introduced that checks the rotation
of each airplane on every frame. This can be optimized in the future to
occur only when the flight plan of an airplane is updated.