feat gh-#1 java sdk and example
Adds a gradle build script to compile the atc api with google's
protobuf gradle plugin. This copies them into the src folder, rather
than accepting the default build folder for ease of copying.
Until jdno publishes this to maven central or a similar public repository,
use `gradle publishToMavenLocal` to generate the corresponding jar.
Its readme is spare, but largely follows the existing rust example.

This likewise adds an example class that connects either synchronously
or asynchronously to a running ATC server and starts the game.

Signed-off-by: Nicholas Prado <nmprado@nzen.ws>