Fix build warnings with tonic and Rust 1.63.0+
Rust 1.63.0 introduced a new lint that checks if `PartialEq` is derived
without also deriving `Eq`. This causes warnings when building the game,
since tonic does not derive `Eq` for the gRPC message types. The current
recommendation is to explicitly allow the lint in the module that
imports the Protocol Buffers.

Read more: https://github.com/hyperium/tonic/issues/1056