Prepare repository for complete rewrite (#251)
After some experimentation and contemplation over the past few weeks,
I have decided to rewrite the game using a different architecture. The
current approach had the following problems:

  - Bevy is still very young and changing a lot. Updating to new
    versions is very work-intensive due to breaking APIs, and the game
    engine is missing a lot of features compared to alternatives.
  - The multi-threaded and non-deterministic nature of Bevy makes it
    difficult for players to develop and test their bots. Given the
    simplicity of the game, a single-threaded game loop with
    deterministic event ordering would probably be better.

The idea is to break the game into three pieces that communicate via two
asynchronous channels with each other:

  1. A simulation engine that runs the game's core loop
  2. An API layer that runs the gRPC server
  3. A UI that allows players to observe the game visually

While a rewrite of the game is risky, I believe that this decision will
unlock many improvements that are now very hard to do.