Reorganize crates
The previous iteration of the game was implemented in a single crate,
but for this iteration it makes more sense to break the game into
individual parts. Namely, the `simulation` crate will implement the
game's core logic and the `api` crate will provide the API layer. This
will make it possible to embed the simulation in other applications,
e.g. to provide a custom user interface for the game.