Refactor state machine in simulation
The state machine has been moved from the simulation itself into a
child. This makes it possible for the simulation to listen for the
command to start a game and then transition its child to the new state.
This wasn't possible before, since the transition between states
consumes the previous state.