Switch game state from a state machine to an enum
The approach with the state machine ended in a dead end, because it is
not possible to store a reference to the state machine without locking
down its type. For example, the simulation cannot initialize the game as
`Game<Ready>` and then change it to `Game<Running>` since that would
change the type of the field.

Instead of a state machine, we are for now going with a simple enum that
represents the state of the simulation. This is less elegant, but easier
to work with.