Create map (#263)
A new struct has been added that represents the map in a running game.
The map has a fixed width and height for the duration of the game, and
fields for its human-readable name and a list of routing nodes.

Maps can be loaded from files, making it easier to iterate on them. For
now, the list of maps has been hard-coded into the program. But
ultimately, it would be nice to allow players to create their own maps
and share them with the community. This will require rewriting parts of
the `MapLoader`, though.