Spawn airplanes (#267)
Airplanes are spawned on a two second timer along the edge of the map.
The implementation makes heavy uses of reference counters to ensure that
the map, airplane spawner, and the airplanes themselves use the same
nodes in the routing grid. While this is not strictly necessary at this
time, it is an important requirement for more dynamic maps in the future
(e.g. nodes that change their restriction status while playing).