Switch to ECS
While implementing the spawner and planning the next features, we
noticed that the current architecture was complex to manage and
difficult to change. Based on the previous success with Bevy's ECS, we
decided to go back to an ECS-based implementation.

Sadly, Bevy's ECS is not yet deterministic (something which might change
with the next release). Based on some quick research, hecs seems to be
the most suitable alternative. It inspired Bevy's ECS, is actively
maintained, and has a simple API that gives us full control over system
execution.