Move airplanes
The system to move airplanes has been migrated to this version of the
game. The algorithm is mostly the same, with slight changes to the inner
loop and used types. Previously, we could rely on Bevy to calculate some
of the 2D geometry. This has been ported to functionality provided by
the geo crate now.