Migrate API layer
The API layer has been copied into the project from the previous
implementation. The services are largely unchanged, but the translation
between internal and external data representations has been completely
rewritten.

Because the lifecycle of some entities has changed, namely the map and
the grid, the tests that had been written for the API don't work
anymore. These will be refactored and added again in a later commit.