Fix panic when generating a flight plan (#288)
The system that generates flight plans crashed in rare instances due to
the following bug:

The system assumes that it will always be able to find at least one
valid next node. Valid is imporant here, since it rules out the last
node that the airplane passed. This is true for all but one case: the
runway. When a plane entered the runway, the only valid path was through
the airport. Because the airport is surrounded by restricted tiles,
however, the system was unable to generate a full-length flight plan.

The fix for this issue is to bail on the flight plan generation. When
the airplane is above the airport, it will either be landed or the go
around procedure will kick in and generate a new flight plan. Both
options are totally valid.