Fix collision detection for airplanes
The collision detection for airplanes was broken due to a misplaced
`break` statement. The detection checked only one other airplane, and if
those didn't collide it would exit the loop prematurely and go to the
next. Instead, the intention was to break after first collision. This
has been fixed, and a test has been added to ensure this stays fixed.